Fix Hammer memory leak
Courtesy of Kerry Davis @ Valve:

"We identified a couple of issues with that version of hammer - it wasn't compiled with the /LARGEADDRESSAWARE flag support to support addresses larger than 2GB, and it has a bug (fixed in our other branches) that causes Hammer to initialize shaders for every material in the tree, regardless of whether it's being used."

This patch fixes the issues described above. There may be minor unintended visual glitches as a side effect.